# oas/rental-api.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Rental v1 REST API",
        "description": "v1 REST API for [Rental](https://github.com/thealamu/rental).",
        "contact": {
            "name": "thealamu",
            "url": "https://github.com/thealamu"
        },
        "license": {
            "name": "MIT",
            "url": "https://spdx.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "servers": [],
...
    "paths": {
...
        "/auth/login": {
            "get": {
                "summary": "Login",
                "description": "Log a user in",
                "operationId": "auth/login",
                "parameters": [
                    {
                        "name": "state_url",
                        "in": "query",
                        "description": "Url to redirect to after a successful log in",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uri",
                            "example": "http://localhost:8080"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "auth successful"
                    }
                }
            }
        }
    }
}